packages/dsw-data-seeder/requirements.txt
certifi==2022.6.15
click==8.1.3
colorama==0.4.5
minio==7.1.10
psycopg2==2.9.3
PyYAML==6.0
tenacity==8.0.1
urllib3==1.26.10
